+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## [unreleased]
 - Add workaround for CSP in Safari < 16 (#6947)
+- Improve add button labels on autotest settings page (#6948)
 
 ## [v2.4.4]
 - Fix websocket connection when logged in with remote auth (#6912)